common/subscriber/pyproject.toml
[project]
name = "subscriber"
version = "0.1.0"
description = ""
authors = [
    {name = "Ian King",email = "ian.antking@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "subscriber", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
